# docs/swagger.yaml
definitions:
...
  disability_system_backend_internal_modules_usuarios_dto.ActualizarUsuarioRequest:
    properties:
      correo:
        maxLength: 150
        type: string
      direccion:
        type: string
      id_rol:
        type: integer
      nombre:
        maxLength: 150
        minLength: 3
        type: string
      numero_celular:
        type: string
    required:
    - correo
    - id_rol
    - nombre
    type: object
  disability_system_backend_internal_modules_usuarios_dto.AsignarRolRequest:
    properties:
      id_rol:
        type: integer
    required:
    - id_rol
    type: object
  disability_system_backend_internal_modules_usuarios_dto.CambiarEstadoUsuarioRequest:
    properties:
      estado:
        type: boolean
    type: object
  disability_system_backend_internal_modules_usuarios_dto.CambiarPasswordRequest:
    properties:
      password_actual:
        type: string
      password_nuevo:
        maxLength: 100
        minLength: 8
        type: string
    required:
    - password_actual
    - password_nuevo
    type: object
  disability_system_backend_internal_modules_usuarios_dto.CrearUsuarioRequest:
    properties:
      correo:
        maxLength: 150
        type: string
      direccion:
        type: string
      id_rol:
        type: integer
      nombre:
        maxLength: 150
        minLength: 3
        type: string
      numero_acudiente:
        type: string
      numero_celular:
        type: string
      numero_documento:
        maxLength: 30
        minLength: 5
        type: string
      password:
        maxLength: 100
        minLength: 8
        type: string
    required:
    - correo
    - id_rol
    - nombre
    - numero_documento
    - password
    type: object
...
info:
  contact: {}
paths:
...
  /usuarios:
    get:
      consumes:
      - application/json
      description: Lista todos los usuarios con paginación
      parameters:
      - default: 1
        description: Número de página
        in: query
        name: page
        type: integer
      - default: 20
        description: Límite de resultados
        in: query
        name: limit
        type: integer
      - description: Filtrar por estado
        in: query
        name: estado
        type: boolean
      - description: Filtrar por rol
        in: query
        name: id_rol
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar usuarios
      tags:
      - usuarios
    post:
      consumes:
      - application/json
      description: Crea un nuevo usuario en el sistema
      parameters:
      - description: Datos del usuario
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/disability_system_backend_internal_modules_usuarios_dto.CrearUsuarioRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Crear usuario
      tags:
      - usuarios
  /usuarios/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina un usuario del sistema (soft delete)
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Eliminar usuario
      tags:
      - usuarios
    get:
      consumes:
      - application/json
      description: Obtiene los detalles de un usuario específico
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener usuario por ID
      tags:
      - usuarios
    put:
      consumes:
      - application/json
      description: Actualiza los datos de un usuario existente
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: integer
      - description: Datos a actualizar
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/disability_system_backend_internal_modules_usuarios_dto.ActualizarUsuarioRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Actualizar usuario
      tags:
      - usuarios
  /usuarios/{id}/estado:
    patch:
      consumes:
      - application/json
      description: Activa o desactiva un usuario
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: integer
      - description: Nuevo estado
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/disability_system_backend_internal_modules_usuarios_dto.CambiarEstadoUsuarioRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Cambiar estado de usuario
      tags:
      - usuarios
  /usuarios/{id}/password:
    post:
      consumes:
      - application/json
      description: Cambia la contraseña del usuario autenticado
      parameters:
      - description: Contraseñas
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/disability_system_backend_internal_modules_usuarios_dto.CambiarPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Cambiar contraseña
      tags:
      - usuarios
  /usuarios/{id}/rol:
    post:
      consumes:
      - application/json
      description: Asigna un rol diferente a un usuario
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: integer
      - description: ID del rol
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/disability_system_backend_internal_modules_usuarios_dto.AsignarRolRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Asignar rol a usuario
      tags:
      - usuarios
swagger: "2.0"